fix tests after refactor
