+++ python/seldon_core/metrics.py
@@ -1,4 +1,4 @@
-import json
+import logging
 import logging
 import os
 from multiprocessing import Manager
@@ -13,8 +13,8 @@ from prometheus_client.core import (
     HistogramMetricFamily,
 )
 from prometheus_client.utils import floatToGoString
-from utils import get_image_name, get_model_name, NONIMPLEMENTED_MSG, ENV_PREDICTOR_LABELS, get_predictior_version, \
-    get_deployment_name, get_predictor_name
+from seldon_core.env_utils import get_predictior_version, get_predictor_name, \
+    get_deployment_name, get_model_name, get_image_name
 
 logger = logging.getLogger(__name__)
